createScheduleをclassに移行
